Move ScaleContext to PietGpuRenderContext
Use a single ScaleContext for the duration of a render context. This
requires a little bit of slightly ugly refactoring to keep the borrow
checker happy, so that the mutable scale context and the mutable state
needed to encode outlines can be separate; if we keep this structure,
it's possible we might clean this up a it.